+++ lib/src/time/calendar/calendar.dart
@@ -3,13 +3,13 @@ library time.calendar;
 import 'package:date/date.dart';
 import 'holiday.dart';
 
-
 /// A holiday calendar
 abstract class Calendar {
   bool isHoliday(Date date);
   HolidayType getHolidayType(Date date);
 
   /// Get the first business day of this month for this calendar.
+  /// TODO: replace this with firstBusinessDateFrom(Date date)
   Date firstBusinessDate(Month month) {
     var candidate = month.startDate;
     while (candidate.isWeekend() || isHoliday(candidate)) {
@@ -17,5 +17,6 @@ abstract class Calendar {
     }
     return candidate;
   }
-}
 
+  /// TODO: add lastBusinessDateBefore(Date date)
+}