+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 2.0.1, released 2021-11-27
+- Performance improvements when calculating if a date is a holiday.  This had
+  cascading beneficial effects in a lot of bucket operations. 
+- Performance improvements for PriceCurve.  When calculating the value for a 
+  given term and bucket no longer force the calculation to hourly curve. 
+  Do that as a last resort only because it's so expensive.  Implement special 
+  treatment for cases when buckets are already in the PriceCurve, etc.
+  These two changes have lead to a 10x speedup in some tests. 
+
 ## 2.0.0, released 2021-05-20
 - Move to null safety (first pass).
 
@@ -12,5 +21,5 @@
 ## 0.5.0, released 2020-07-05
 - Snapshot of work
 
-## 0.0.1
+## 0.0.1, released 2015-09-27
 - Initial version, created by Stagehand